requirements.txt
astroid==2.4.2
autopep8==1.5.4
certifi==2020.12.5
chardet==3.0.4
idna==2.10
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
pycodestyle==2.6.0
pyflakes==2.2.0
pylint==2.6.0
PySocks==1.7.1
python-dotenv==0.15.0
requests==2.25.0
requests-random-user-agent==2020.12.4
six==1.15.0
stem==1.8.0
toml==0.10.2
urllib3==1.26.2
wrapt==1.12.1
